auditor/requirements.txt
SQLAlchemy
grpcio
grpcio-health-checking
protobuf
webauthn
python-dotenv
pyjwt
eth_keys
eth_account
web3
pymysql
requests
python-bitcoinlib
sqlalchemy-stubs
sqlalchemy-utils
pylint
black
pytz
py-cid
isort
git+https://github.com/ravirahman/petlib@3b28c97955655671fbc167e007ab6f7ea438801b#egg=petlib
../protobufs/python
../common
